requirements.txt
numpy
Pillow
kaggle
